ci: fixed CodeQL analysis workflow and building with 64-bits (#92)
* ci: fixed CodeQL code build stage
* ci: added workflow deploy for codeql-fix branch
* game_shared: utlmemory.h pointer type fix
* common: areanode.h pointer type fix
* client: entity.cpp fixed pointer type
* server: cbase.h fixed CBaseEntity::SetParent ambigious call
* server: physint.h fixed invalid pointer type
* server: fixed ambigious SetParent call in CCrossbowBolt::OnChangeLevel
* server: fixed items.cpp ambigious call
* server: fixed rest of ambigious SetParent calls

+++ common/areanode.h
@@ -16,6 +16,9 @@ GNU General Public License for more details.
 #ifndef AREANODE_H
 #define AREANODE_H
 
+#include "const.h"
+#include <stddef.h>
+
 typedef struct areanode_s
 {
 	int		axis;		// -1 = leaf node
@@ -26,7 +29,7 @@ typedef struct areanode_s
 	link_t		portal_edicts;
 } areanode_t;
 
-#define STRUCT_FROM_LINK( l, t, m )		((t *)((byte *)l - (int)&(((t *)0)->m)))
+#define STRUCT_FROM_LINK( l, t, m )		((t *)((byte *)l - (size_t)&(((t *)0)->m)))
 #define EDICT_FROM_AREA( l )			STRUCT_FROM_LINK( l, edict_t, area )
 #define FACET_FROM_AREA( l )			STRUCT_FROM_LINK( l, mfacet_t, area )
